test: share app/store harness lifecycle between HTTP and socket suites
Move loadAppModule/loadJsonStore/loadGatewayModule/SharedTestContext and the
beforeAll/afterEach/afterAll wiring into test/lib/shared/harness.ts as a
generic useHarnessLifecycle(createHarnessInstance, onAfterEach?) helper. Both
useHttpHarness() and useSocketHarness() now only define their own
createHarnessInstance() (Express+supertest vs. Express+Socket.IO) and harness
shape; HTTP keeps its serialPorts mock and per-test reset as an onAfterEach
callback, Socket.IO setup stays local to socket/harness.ts.
