+++ test/lib/http/harness.ts
@@ -1,19 +1,17 @@
 import { createServer, type Server as HttpServer } from 'node:http'
 import type { Express } from 'express'
 import supertest, { type Test as SupertestTest } from 'supertest'
-import { vi, beforeAll, afterEach, afterAll } from 'vitest'
-import { cleanupTestEnv, ensureTestEnv } from '../shared/env.ts'
+import { vi } from 'vitest'
 import type { FakeGateway } from '../shared/fakes.ts'
-import type * as AppModuleNamespace from '#api/app.ts'
-import type * as JsonStoreModuleNamespace from '#api/lib/jsonStore.ts'
-import type * as StoreConfigModuleNamespace from '#api/config/store.ts'
-import type * as GatewayModuleNamespace from '#api/lib/Gateway.ts'
 import type { Driver } from 'zwave-js'
+import {
+	useHarnessLifecycle,
+	type GatewayModule,
+	type JsonStoreModule,
+	type SharedTestContext,
+	type StoreConfigModule,
+} from '../shared/harness.ts'
 
-type AppModule = typeof AppModuleNamespace
-type JsonStoreModule = typeof JsonStoreModuleNamespace
-type StoreConfigModule = typeof StoreConfigModuleNamespace
-type GatewayModule = typeof GatewayModuleNamespace
 type RealGateway = InstanceType<GatewayModule['default']>
 
 type SerialPortsEnumerator = typeof Driver.enumerateSerialPorts
@@ -21,33 +19,6 @@ type SerialPortsEnumeratorMock = ReturnType<typeof vi.fn<SerialPortsEnumerator>>
 
 function createSerialPortsEnumeratorMock(): SerialPortsEnumeratorMock {
 	return vi.fn(() => Promise.resolve<string[]>([]))
-}
-
-let appModulePromise: Promise<AppModule> | undefined
-let jsonStoreModulePromise: Promise<JsonStoreModule> | undefined
-let storeConfigModulePromise: Promise<StoreConfigModule> | undefined
-let gatewayModulePromise: Promise<GatewayModule> | undefined
-
-export async function loadAppModule(): Promise<AppModule> {
-	ensureTestEnv()
-	appModulePromise ??= import('#api/app.ts')
-	return appModulePromise
-}
-
-export async function loadJsonStore() {
-	ensureTestEnv()
-	jsonStoreModulePromise ??= import('#api/lib/jsonStore.ts')
-	storeConfigModulePromise ??= import('#api/config/store.ts')
-	const [jsonStoreMod, storeMod] = await Promise.all([
-		jsonStoreModulePromise,
-		storeConfigModulePromise,
-	])
-	return { jsonStore: jsonStoreMod.default, store: storeMod.default }
-}
-
-async function loadGatewayModule(): Promise<GatewayModule> {
-	gatewayModulePromise ??= import('#api/lib/Gateway.ts')
-	return gatewayModulePromise
 }
 
 export interface HttpHarness {
@@ -74,24 +45,6 @@ export interface HttpHarnessOptions {
 	gateway?: FakeGateway
 	restarting?: boolean
 	enumerateSerialPorts?: SerialPortsEnumeratorMock
-}
-
-interface SharedTestContext {
-	createApp: AppModule['createApp']
-	jsonStore: JsonStoreModule['default']
-	store: StoreConfigModule['default']
-	closeWatchers: GatewayModule['closeWatchers']
-}
-
-async function createSharedTestContext(): Promise<SharedTestContext> {
-	const [{ createApp }, { jsonStore, store }, { closeWatchers }] =
-		await Promise.all([
-			loadAppModule(),
-			loadJsonStore(),
-			loadGatewayModule(),
-		])
-	await jsonStore.init(store)
-	return { createApp, jsonStore, store, closeWatchers }
 }
 
 async function createHarnessInstance(
@@ -138,38 +91,5 @@ async function createHarnessInstance(
 export function useHttpHarness(): (
 	options?: HttpHarnessOptions,
 ) => Promise<HttpHarness> {
-	let shared: SharedTestContext | undefined
-	let current: (HttpHarness & { closeInstance(): Promise<void> }) | undefined
-
-	beforeAll(async () => {
-		shared = await createSharedTestContext()
-	})
-
-	afterEach(async () => {
-		if (current) {
-			await current.closeInstance()
-			current = undefined
-		}
-	})
-
-	afterAll(() => {
-		if (!shared) return
-		shared.closeWatchers()
-		for (const key of Object.keys(shared.jsonStore.store)) {
-			delete shared.jsonStore.store[key]
-		}
-		cleanupTestEnv()
-	})
-
-	return async function getHarness(
-		options: HttpHarnessOptions = {},
-	): Promise<HttpHarness> {
-		if (!shared) {
-			throw new Error(
-				'useHttpHarness(): beforeAll has not run yet, call the accessor from within a test',
-			)
-		}
-		current ??= await createHarnessInstance(shared, options)
-		return current
-	}
+	return useHarnessLifecycle(createHarnessInstance)
 }
